added aliases, simplified component file structure
