+++ App.js
@@ -13,12 +13,12 @@ import EntryPoint from './App/screens/EntryPoint/index'
 // import Home from './App/screens/Home/index'
 // import ListContacts from './App/screens/ListContacts/index'
 // import ListCredentials from './App/screens/ListCredentials/index'
-import Navbar from './App/components/Navbar/index'
+import Navbar from './App/components/Navbar'
 import PinCreate from './App/screens/PinCreate/index'
 import PinEnter from './App/screens/PinEnter/index'
 // import Settings from './App/screens/Settings/index'
 import SetupWizard from './App/screens/SetupWizard/index'
-import Terms from './App/components/Terms/index'
+import Terms from './App/components/Terms'
 import Workflow from './App/screens/Workflow/index'
 
 import TabNavigator from './App/navigators/TabNavigator'
