+++ App/components/AppHeader.tsx
@@ -1,12 +1,11 @@
 import React from 'react'
 
-import { Image, Text, TouchableOpacity, View } from 'react-native'
+import { Image, Text, TouchableOpacity, View, StyleSheet } from 'react-native'
 
 import { useHistory } from 'react-router-native'
 
-import Images from '../../../assets/images'
-import AppStyles from '../../../assets/styles'
-import Styles from './styles'
+import Images from '../../assets/images'
+import AppStyles from '../../assets/styles'
 
 interface AppHeaderProps {
   headerText?: string
@@ -40,3 +39,22 @@ function AppHeader(props: AppHeaderProps) {
 }
 
 export default AppHeader
+
+const Styles = StyleSheet.create({
+  header: {
+    height: '20%',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  headerView: {
+    flex: 1,
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingHorizontal: 10,
+  },
+  textBorder: {
+    borderBottomWidth: 3,
+    borderColor: '#1b2624',
+  },
+})
