--- App/components/AppHeader/styles.js
@@ -1,24 +1,0 @@
-import React from 'react'
-
-import { StyleSheet } from 'react-native'
-
-const Styles = StyleSheet.create({
-  header: {
-    height: '20%',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  headerView: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: 10,
-  },
-  textBorder: {
-    borderBottomWidth: 3,
-    borderColor: '#1b2624',
-  },
-})
-
-export default Styles
