--- App/components/AppHeaderLarge/styles.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import { StyleSheet } from 'react-native'
-
-const Styles = StyleSheet.create({
-  headerLarge: {
-    height: '40%',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-})
-
-export default Styles
