+++ App/components/CurrentContact.tsx
@@ -1,13 +1,12 @@
 import React from 'react'
 
-import { Image, Text, TouchableOpacity, View } from 'react-native'
+import { Image, Text, TouchableOpacity, View, StyleSheet } from 'react-native'
 
 import { useHistory } from 'react-router-native'
 
-import AppStyles from '../../../assets/styles'
-import Images from '../../../assets/images'
-import Styles from './styles'
-import { IContact } from '../../../types'
+import AppStyles from '../../assets/styles'
+import Images from '../../assets/images'
+import { IContact } from '../../types'
 
 interface ICurrentContact {
   setViewContact: (toggle: boolean) => void
@@ -64,3 +63,25 @@ function CurrentContact(props: ICurrentContact) {
 }
 
 export default CurrentContact
+
+const styles = StyleSheet.create({
+  backbutton: {
+    marginBottom: 30,
+  },
+  tableItem: {
+    paddingLeft: 30,
+    display: 'flex',
+    alignItems: 'center',
+  },
+  tableSubItem: {
+    height: 50,
+  },
+  credView: {
+    alignItems: 'center',
+    backgroundColor: '#ddd',
+    padding: 12,
+    borderTopLeftRadius: 50,
+    borderTopRightRadius: 50,
+    height: '100%',
+  },
+})
