+++ App/components/CurrentCredential.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 
-import { Image, Text, TouchableOpacity, View } from 'react-native'
+import { Image, Text, TouchableOpacity, View, StyleSheet } from 'react-native'
 
 import { useHistory } from 'react-router-native'
 
-import AppStyles from '../../../assets/styles'
-import Images from '../../../assets/images'
-import Styles from './styles'
-import { ICredential } from '../../../types'
+import AppStyles from '../../assets/styles'
+import Images from '../../assets/images'
+import Styles from './CurrentCredential/styles'
+import { ICredential } from '../../types'
 
 interface ICurrentCredential {
   credential: ICredential
@@ -67,3 +67,25 @@ function CurrentCredential(props) {
 }
 
 export default CurrentCredential
+
+const styles = StyleSheet.create({
+  backbutton: {
+    marginBottom: 30,
+  },
+  tableItem: {
+    paddingLeft: 30,
+    display: 'flex',
+    alignItems: 'center',
+  },
+  tableSubItem: {
+    height: 50,
+  },
+  credView: {
+    alignItems: 'center',
+    backgroundColor: '#ddd',
+    padding: 12,
+    borderTopLeftRadius: 50,
+    borderTopRightRadius: 50,
+    height: '100%',
+  },
+})
