--- App/components/CurrentCredential/styles.js
@@ -1,27 +1,0 @@
-import React from 'react'
-
-import { StyleSheet } from 'react-native'
-
-const styles = StyleSheet.create({
-  backbutton: {
-    marginBottom: 30,
-  },
-  tableItem: {
-    paddingLeft: 30,
-    display: 'flex',
-    alignItems: 'center',
-  },
-  tableSubItem: {
-    height: 50,
-  },
-  credView: {
-    alignItems: 'center',
-    backgroundColor: '#ddd',
-    padding: 12,
-    borderTopLeftRadius: 50,
-    borderTopRightRadius: 50,
-    height: '100%',
-  },
-})
-
-export default styles
