+++ App/components/Navbar.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 
-import { Image, Text, TouchableOpacity, View } from 'react-native'
+import { Image, Text, TouchableOpacity, View, StyleSheet } from 'react-native'
 
 import { Link } from 'react-router-native'
 
-import Images from '../../../assets/images'
-import AppStyles from '../../../assets/styles'
-import Styles from './styles'
+import Images from '../../assets/images'
+import AppStyles from '../../assets/styles'
+import Styles from './Navbar/styles'
 
 interface INavbar {
   authenticated: boolean
@@ -43,3 +43,32 @@ function Navbar(props: INavbar) {
 }
 
 export default Navbar
+
+const styles = StyleSheet.create({
+  navView: {
+    width: '101%',
+    top: -8,
+    marginLeft: '-0.5%',
+    height: '100%',
+    padding: 20,
+    paddingTop: 14,
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-around',
+    bottom: 0,
+    borderTopWidth: 2,
+    borderRightWidth: 0.1,
+    borderLeftWidth: 0.1,
+    borderColor: 'rgba(0,0,0,0.2)',
+    borderBottomWidth: 0,
+    borderTopLeftRadius: 16,
+    borderTopRightRadius: 16,
+  },
+  navButton: {
+    alignItems: 'center',
+    width: '20%',
+  },
+  textSmall: {
+    fontSize: 10,
+  },
+})
