+++ App/components/PAButton.tsx
@@ -1,7 +1,12 @@
 import React from 'react'
 import { TouchableOpacity, Text, StyleSheet } from 'react-native'
 
-function PAButton({ title, onPress, ...otherProps }) {
+interface Props {
+  title: string
+  onPress: () => void
+}
+
+const PAButton: React.FC<Props> = ({ title, onPress, ...otherProps }) => {
   return (
     <TouchableOpacity onPress={onPress} style={styles.button} {...otherProps}>
       <Text style={styles.text}>{title}</Text>
@@ -14,10 +19,10 @@ export default PAButton
 const styles = StyleSheet.create({
   button: {
     width: '85%',
-    borderRadius: 10,
+    borderRadius: 5,
     backgroundColor: '#35823f',
     alignItems: 'center',
-    padding: 20,
+    padding: 10,
     margin: 10,
   },
   text: {
