+++ App/screens/EntryPoint/index.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-native'
 import * as Keychain from 'react-native-keychain'
 import AsyncStorage from '@react-native-community/async-storage'
 
-import LoadingOverlay from '../../components/LoadingOverlay/index'
+import LoadingOverlay from '../../components/LoadingOverlay'
 
 import { ErrorsContext } from '../../contexts/Errors/index'
