+++ package.json
@@ -55,6 +55,7 @@
     "@typescript-eslint/eslint-plugin": "^4.20.0",
     "@typescript-eslint/parser": "^4.20.0",
     "babel-jest": "^26.6.3",
+    "babel-plugin-module-resolver": "^4.1.0",
     "commitlint": "^11.0.0",
     "eslint": "^7.14.0",
     "eslint-plugin-prettier": "^3.3.1",
